server/requirements.txt
flask
flask_cors
flask_mysqldb
bcrypt
pymysql
pillow
